refactor: 레스토랑 추가 시 수행하는 함수 inputForm으로 이동
